UPDATED: top-regions with stacked architecture :fire:
