+++ lib/ui/pages/views/home_view/home_view.dart
@@ -1,6 +1,7 @@
 import 'package:Covid19_PH/ui/pages/views/home_view/home_view_model.dart';
 import 'package:Covid19_PH/ui/partials/views/summary_view/summary_view.dart';
 import 'package:Covid19_PH/ui/partials/views/timeline_view/timeline_view.dart';
+import 'package:Covid19_PH/ui/partials/views/top_regions_view/top_regions_view.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
@@ -24,7 +25,7 @@ class HomeView extends StatelessWidget {
                 SizedBox(height: 30.0),
                 TimelineView(),
                 SizedBox(height: 30.0),
-                // TopRegionsWidget(),
+                TopRegionsView(),
                 SizedBox(height: 30.0),
               ],
             ),
