+++ lib/ui/partials/views/top_regions_view/concretes/top_regions_card.dart
